Remove WordSeqProvider and DocProvider
